Update FloatingTips added 'above' position
Added 'above' position which works as 'inside' but without injecting the
tip inside the element it is bound to. It is usefull in those cases
where the element wich has the tip has a defined widht, height and/or
overflow set to hidden (to avoid the tip to appear cropped).