[FEATURE] Allow tip hiding on its outside click
This adds an option `hideOnTipOutsideClick` which will
hide the Tip if the document is clicked somewhere outside
the ToolTip resp. its initial trigger element.

Apparently this only makes sense if no contradicting
`hideOn` option is set, as it would be the case in most
`mouseleave` cases because there would be no opportunity
to click outside if the Tip is just hiding anyways.

Fixes #21